Fix IOHelper.postRequest to not use static 30ms connect timeout
Change MapLayer constructor to allow Integer (for MyBatis), change type of mapLayerId to int to reflect the DB schema

+++ service-spatineo-monitor/src/main/java/org/oskari/spatineo/monitor/UpdateBackendStatusJob.java
@@ -130,7 +130,7 @@ public class UpdateBackendStatusJob {
         return i < 0 ? layerName : layerName.substring(0, i);
     }
 
-    private static BackendStatus getStatus(long mapLayerId, Meter meter) {
+    private static BackendStatus getStatus(int mapLayerId, Meter meter) {
         Indicator indicator = meter.getIndicator();
         String statusMessage = indicator.getStatus();
         Status status = Status.getEnumByNewAPI(statusMessage);
