+++ service-spatineo-monitor/src/main/java/org/oskari/spatineo/monitor/backendstatus/BackendStatus.java
@@ -2,19 +2,19 @@ package org.oskari.spatineo.monitor.backendstatus;
 
 public class BackendStatus {
 
-    private final long mapLayerId;
+    private final int mapLayerId;
     private final String status;
     private final String statusMessage;
     private final String infoUrl;
 
-    public BackendStatus(long mapLayerId, String status, String statusMessage, String infoUrl) {
+    public BackendStatus(int mapLayerId, String status, String statusMessage, String infoUrl) {
         this.mapLayerId = mapLayerId;
         this.status = status;
         this.statusMessage = statusMessage;
         this.infoUrl = infoUrl;
     }
 
-    public long getMapLayerId() {
+    public int getMapLayerId() {
         return mapLayerId;
     }
